api/requirements.txt
suds-jurko>=0.6

